requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
flask==2.3.3
Flask-Login==0.6.2
Flask-Silk==0.2
flask-sqlalchemy==3.0.5
future==0.18.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.20
typing-extensions==4.7.1
urllib3==2.0.4
werkzeug==2.3.7
zipp==3.16.2
